CI: Fix the tag.yml does not release issue
When tag.yml is triggered on branch master, the git reference has not
been at a tag. So, when the main.yml is used by tag.yml, the git
refernece has not been at a tag as well. This makes the release
condition check startsWith(github.ref, 'refs/tags/') becomes failed.

Therefore, pass argument is_at_tag into main.yml to know git HEAD is at
a tag, or not.

https://phabricator.endlessm.com/T33610