CI: Pass the latest tag name from tag.yml to main.yml
The release step with softprops/action-gh-release@v1 [1] uploading the
artifact to the release page needs the tag_name. The tag_name defaults
to github.ref.

However, the github.ref is not at a tag when the tag.yml is triggered on
master branch, even CI has created a tag. For example, the schedule
event. This leads softprops/action-gh-release@v1 fails to upload the
artifact.

This commit fixes the issue by passing the tag's name as the inputs of
softprops/action-gh-release@v1.

[1] https://github.com/softprops/action-gh-release

https://phabricator.endlessm.com/T33610