electron: Set KOLIBRI_PROJECT env by detecting app's directory
Endless Key (Windows) is only delivered as two types: APPX/MSIX from
Microsoft store, or the zip/image format which might be in an USB key
currently.

We would like to know it runs in which environment. So, only set the
KOLIBRI_PROJECT env as 'endless-key-windows' as the metrics ID when it's
directory name includes APPX/MSIX pattern. Others fall back as
'endless-key-windows-standalone'.

https://phabricator.endlessm.com/T33665