Actually set DJANGO_SETTINGS_MODULE
If you import anything whatsoever from the 'kolibri' namespace in Python
code, a small number of environment variables are set to a default value
if they are not already set. Specifically, kolibri/__init__.py has a
top-level call to kolibri.utils.env.set_env(). DJANGO_SETTINGS_MODULE is
one of those variables.

We want to use kolibri_tools.endless_key_settings, which takes
kolibri.deployment.default.settings.base, the default
DJANGO_SETTINGS_MODULE, and adds a single middleware,
kolibri_tools.middleware.AlwaysAuthenticatedMiddleware. This middleware
causes unauthenticated requests to be treated as logged-in requests from
the "endless" user.

Previously, we attempted to set DJANGO_SETTINGS_MODULE midway through
the entrypoint, main.pyw, if it was not already set. However one of the
first things main.pyw does is import kolibri_tools.utils, which in turn
imports some stuff from kolibri.plugins, and boom! kolibri/__init__.py
is executed, which sets DJANGO_SETTINGS_MODULE if not already set. So
the code in main.pyw has no effect. As a result,
AlwaysAuthenticatedMiddleware is not used, and functionality that
depends on it being present (such as downloading additional items)
fails.

I can see two ways to fix this:

- In main.pyw, set this environment variable earlier, before
  (transitively) importing anything from the 'kolibri' module namespace.
- Set the environment variable in the node.js/electron launcher
  application instead, before Python is spawned.

I chose the latter for consistency: KOLIBRI_HOME is one of the other
variables that gets defaulted in this way, and we set that variable in
the electron launcher too.

Fixes https://github.com/endlessm/endless-key-app/issues/128