Fix KOLIBRI_HOME as real path of %APPDATA%\kolibri-electron\endless-key
The %APPDATA% is a mapped virtual path for APPX/MSIX package on Windows.
However, the cookies' storage path created by Windows' native webview
invoked by pywebview is not the mapped virtual path.

It is a knnown issue mentioned in Python document "Redirection of local
data, registry, and temporary paths" [1]

So, have to get the real path of the created KOLIBRI_HOME folder and
reset variable KOLIBRI_HOME to the real path of the folder.

[1]: https://docs.python.org/3.11/using/windows.html#redirection-of-local-data-registry-and-temporary-paths